test-available/mem0-integration/mem0-server-local/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.4
mem0ai>=0.1.48
python-dotenv==1.0.1
faiss-cpu>=1.7.4
pyinstaller>=6.3.0
